+++ .prettierrc.cjs
@@ -1,11 +1,8 @@
 module.exports = {
   pluginSearchDirs: false,
-  plugins: [
-    require.resolve('prettier-plugin-organize-imports'),
-    require.resolve('prettier-plugin-packagejson'),
-  ],
-  printWidth: 160,
-  proseWrap: 'never',
+  plugins: [require.resolve('prettier-plugin-organize-imports'), require.resolve('prettier-plugin-packagejson')],
+  printWidth: 120,
+  proseWrap: 'preserve',
   singleQuote: true,
   trailingComma: 'all',
   overrides: [