+++ packages/banana-react/src/index.ts
@@ -4,10 +4,26 @@ import { Collapse } from './collapse';
 import { Countdown } from './countdown';
 import { Divider } from './divider';
 import { Dropdown } from './dropdown';
+import { Input } from './input';
 import { Menu } from './menu';
 import { MenuItem } from './menu-item';
 import { Overlay } from './overlay';
 import { Popup } from './popup';
 import { Progress } from './progress';
+import { Rating } from './rating';
 
-export { Button, Carousel, Collapse, Countdown, Divider, Dropdown, Menu, MenuItem, Overlay, Popup, Progress };
+export {
+  Button,
+  Carousel,
+  Collapse,
+  Countdown,
+  Divider,
+  Dropdown,
+  Input,
+  Menu,
+  MenuItem,
+  Overlay,
+  Popup,
+  Progress,
+  Rating,
+};