+++ packages/banana-react/src/input/index.ts
@@ -1,0 +1,12 @@
+import { createComponent } from '@lit-labs/react';
+import { BInput } from 'banana-ui';
+import * as React from 'react';
+
+export const Input = createComponent({
+  tagName: 'b-input',
+  react: React,
+  elementClass: BInput,
+  events: {
+    onChange: 'change',
+  },
+});