+++ packages/banana-react/src/rating/index.ts
@@ -1,0 +1,9 @@
+import { createComponent } from '@lit-labs/react';
+import { BRating } from 'banana-ui';
+import * as React from 'react';
+
+export const Rating = createComponent({
+  tagName: 'b-rating',
+  react: React,
+  elementClass: BRating,
+});