+++ packages/banana/src/index.ts
@@ -4,10 +4,26 @@ import BCollapse from './collapse';
 import BCountdown from './countdown';
 import BDivider from './divider';
 import BDropdown from './dropdown';
+import BInput from './input';
 import BMenu from './menu';
 import BMenuItem from './menu-item';
 import BOverlay from './overlay';
 import BPopup from './popup';
 import BProgress from './progress';
+import BRating from './rating';
 
-export { BButton, BCarousel, BCollapse, BCountdown, BDivider, BDropdown, BMenu, BMenuItem, BOverlay, BPopup, BProgress };
+export {
+  BButton,
+  BCarousel,
+  BCollapse,
+  BCountdown,
+  BDivider,
+  BDropdown,
+  BInput,
+  BMenu,
+  BMenuItem,
+  BOverlay,
+  BPopup,
+  BProgress,
+  BRating,
+};