+++ packages/banana/CHANGELOG.md
@@ -1,5 +1,41 @@
 # banana-ui
 
+## 1.10.0
+
+### Minor Changes
+
+- 34549ad: Beta version for form, input, rating.
+
+### Patch Changes
+
+- 37ac10a: Fixed a bug which cause defaultValue paramter not working on rating component.
+- 375f8ff: Rating: Keep internal value and value property synchronized.
+- Added controlled parameter to input component.
+
+## 1.10.0-beta.3
+
+### Patch Changes
+
+- Added controlled parameter to input component.
+
+## 1.10.0-beta.2
+
+### Patch Changes
+
+- Rating: Keep internal value and value property synchronized.
+
+## 1.10.0-beta.1
+
+### Patch Changes
+
+- Fixed a bug which cause defaultValue paramter not working on rating component.
+
+## 1.10.0-beta.0
+
+### Minor Changes
+
+- Beta version for form, input, rating.
+
 ## 1.9.3
 
 ### Patch Changes